Switch out old search expression parser and replace with lark
Remove PyPEG2 dependency
